siglent-sds: improve not supported trigger message
When connecting to a Siglent SDS1204X-E with a SERIAL (SPI) trigger
enabled, it was observed that sigrok-cli failed with no clear indication
as to why.  The problem was debugged down to the siglent-sds driver only
supports simple triggers at the moment, not protocol analyzer triggers.

Print a helpful error message so others facing the same dilema will
quickly deduce they need to change their enabled trigger before
connecting with sigrok.

In the future, it may be better to support an "unknown" or
"device-only" trigger.