siglent-sds: correct 4 channel trigger source list
When connecting a 4 channel Siglent SDS 1204X-E scope to pulseview, it
was observed that only analog channels CH1 and CH2 showed up in the list
of available trigger sources.

Fix the list of trigger sources by building at probe time.  This allows
us to build the list based on the number of available channels with the
given hardware.

Tested on a Siglent SDS 1204X-E scope.